+++ tests/application_tests/src/test_api.py
@@ -10,6 +10,6 @@ class ApiTest(unittest.TestCase):
 
     def test_documentation(self):
         """Test that the documentation API is available."""
-        apis = requests.get("http://www:8080/api", timeout=10).json().keys()
+        apis = requests.get("http://www:8080/api/v3/docs", timeout=10).json().keys()
         self.assertIn("/api/internal/login", apis)
         self.assertIn("/api/v3/login", apis)